cleanup: conversion not needed

+++ src/main/kotlin/com/github/K0zka/kerub/data/ispn/AbstractAssetDao.kt
@@ -24,7 +24,7 @@ abstract class AbstractAssetDao<T : Asset>(
 				listOf()
 			} else {
 				cache.queryBuilder(getEntityClass().kotlin)
-						.startOffset(start.toLong())
+						.startOffset(start)
 						.maxResults(limit)
 						.having(field).like("%$value%")
 						.and()
